Localize letters and create them procedurally
Adds localization tables for each separate part of a letter. For the intro, wish and deeds there are three options each, for the three yet supported idioms (en, pt-br and es).
Adds script to select randomly a key for each table and concatenate them for a UI text.
Adds an editor button to create a new letter each time it's pressed.
Adds mybox package to project.
Letters must be attached to a player (or any person, whatsoever) to know their name and gender.
For now, supports Male and Female, but non-binary options shall be added later (TODO: learn gender neutral expressions in spanish for this).