Refactors the code and also restructures some responsibilities
Refactors the majority of the code to better preserver encapsulation and distribute responsibilities outside the game manager. Restructures how the letter are handled, making the letters control their amount and call the game ending. Also restructures the ending so that a separate script handles the score and what ending to show. Restructures the score so that the score bar handles itself. Changes some events to decouple the game manager and other classes.